Nod Unity SDK Version 1.0.1
Adds Mac support but only for Unity 5.x Beta 18 or greater.  We wont be
able to build a Unity 4.x plugin because we rely on ARC which Apple only
maintains for 64 bit targets and Unity 4.x is 32bit only.

Added a support folder with specialized AndroidManifest.xml's to assist
in compiling to Samsung GearVR, and Google cardboard targets.